Revert "feat: release 4.0.0 (#812) (#857)"
This reverts commit 1a11745fe96e17823f98af455a3e0d224aa26baa.
