+++ docs/useSelect/usage.mdx
@@ -1,0 +1,80 @@
+---
+name: Usage
+menu: useSelect
+route: /hooks/use-select/usage
+---
+
+import {Playground} from 'docz'
+import {useSelect} from '../../src'
+import {items, menuStyles} from './utils.json'
+
+# useSelect
+
+The `useSelect` hook provides functionality and accessibility to a dropdown that should act as a HTML `<select>`, as described by the corresponding [ARIA docs](https://www.w3.org/TR/wai-aria-practices/examples/listbox/listbox-collapsible.html).
+
+User should be allowed to perform a selection after opening the dropdown's menu. Only one item may be selected at a time. The widget should be built using a text label (optional but recommended), a button that toggles the menu, the menu itself and the collection of items that belong in the menu.
+
+## Building the UI
+
+A custom `<select>` element can be created with HTML elements such as: `<label>`, `<ul>`, `<li>` and `<button>`. Using other HTML elements to create a custom `<select>` is useful for the custom styling of the widget, since the `<select>` is notoriously difficult to be styled: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/select
+
+In the example below, we use the `useSelect` hook and destructure from its result the getter props and state variables. These are used in the following way:
+
+| Returned prop          | Element                | Comments                                                                  |
+| ---------------------- | ---------------------- | ------------------------------------------------------------------------- |
+| `getLabelProps`        | `<label>`              | Adds an `id` attribute to be used for `menu` and `triggerButton`          |
+| `getToggleButtonProps` | `<button>`             | Controls the open state of the list.                                      |
+| `getMenuProps`         | `<ul>`                 | Makes list focusable, adds ARIA attributes and event handlers.            |
+| `getItemProps`         | `<li>`                 | Called with `index` and `item`, adds ARIA attributes and event listeners. |
+| `isOpen`               | before `<li>` elements | Only when it's true we render the `<li>` elements.                        |
+| `highlightedIndex`     | highlighted `<li>`     | Used to style the highlighted item.                                       |
+| `selectedItem`         | `<button>`             | Used to render text equivalent of selected item on the button.            |
+
+To build the UI with component libraries such as MaterialUI for example, go to [UI Libraries](/hooks/use-select/ui-libraries).
+
+## Playground
+
+<Playground style={{height: '160px'}}>
+  {() => {
+    const DropdownSelect = () => {
+      const {
+        isOpen,
+        selectedItem,
+        getToggleButtonProps,
+        getLabelProps,
+        getMenuProps,
+        highlightedIndex,
+        getItemProps,
+      } = useSelect({items})
+      return (
+        <div>
+          <label {...getLabelProps()}>Choose an element:</label>
+          <button {...getToggleButtonProps()}>
+            {selectedItem || 'Elements'}
+          </button>
+          <ul {...getMenuProps()} style={menuStyles}>
+            {isOpen &&
+              items.map((item, index) => (
+                <li
+                  style={
+                    highlightedIndex === index
+                      ? {backgroundColor: '#bde4ff'}
+                      : {}
+                  }
+                  key={`${item}${index}`}
+                  {...getItemProps({item, index})}
+                >
+                  {item}
+                </li>
+              ))}
+          </ul>
+        </div>
+      )
+    }
+    return <DropdownSelect />
+  }}
+</Playground>
+
+## CodeSandbox
+
+Codesandbox equivalent is at [here](https://codesandbox.io/s/useselect-usage-53qfj).
