+++ docs/useSelect/utils.json
@@ -1,0 +1,40 @@
+{
+  "items": [
+    "Neptunium",
+    "Plutonium",
+    "Americium",
+    "Curium",
+    "Berkelium",
+    "Californium",
+    "Einsteinium",
+    "Fermium",
+    "Mendelevium",
+    "Nobelium",
+    "Lawrencium",
+    "Rutherfordium",
+    "Dubnium",
+    "Seaborgium",
+    "Bohrium",
+    "Hassium",
+    "Meitnerium",
+    "Darmstadtium",
+    "Roentgenium",
+    "Copernicium",
+    "Nihonium",
+    "Flerovium",
+    "Moscovium",
+    "Livermorium",
+    "Tennessine",
+    "Oganesson"
+  ],
+  "menuStyles": {
+    "maxHeight": "180px",
+    "overflowY": "auto",
+    "width": "200px",
+    "position": "absolute",
+    "margin": 0,
+    "borderTop": 0,
+    "background": "white",
+    "zIndex": 1000
+  }
+}
