+++ doczrc.js
@@ -1,5 +1,5 @@
 export default {
   port: 6006,
   files: './docs/**/*.{md,markdown,mdx}',
-  menu: ['Home', {name: 'Hooks', menu: ['useSelect', 'useCombobox']}, 'Tests'],
+  menu: ['Home', {name: 'useSelect', menu: ['Usage', 'UI Libraries']}, 'Tests'],
 }
