+++ package.json
@@ -73,8 +73,6 @@
   },
   "devDependencies": {
     "@babel/helpers": "^7.4.4",
-    "@material-ui/core": "^4.7.0",
-    "@material-ui/icons": "^4.5.1",
     "@testing-library/cypress": "^5.0.2",
     "@testing-library/jest-dom": "^4.1.2",
     "@testing-library/preact": "^1.0.2",
