+++ src/hooks/index.js
@@ -1,2 +1,1 @@
 export {default as useSelect} from './useSelect'
-export {default as useCombobox} from './useCombobox'
