--- src/hooks/useCombobox/__tests__/utils.test.js
@@ -1,9 +1,0 @@
-import reducer from '../reducer'
-
-describe('utils', () => {
-  test('reducer throws error if called without proper action type', () => {
-    expect(() => {
-      reducer({}, {type: 'super-bogus'})
-    }).toThrowError('Reducer called without proper action type.')
-  })
-})
