--- src/hooks/useCombobox/testUtils.js
@@ -1,107 +1,0 @@
-import React from 'react'
-import {render} from '@testing-library/react'
-import {renderHook} from '@testing-library/react-hooks'
-import {useId, defaultProps} from '../utils'
-import {getElementIds} from './utils'
-import useCombobox from '.'
-
-jest.mock('../utils', () => {
-  const module = require.requireActual('../utils')
-
-  module.useId = () => 'test-id'
-
-  return module
-})
-
-const items = [
-  'Neptunium',
-  'Plutonium',
-  'Americium',
-  'Curium',
-  'Berkelium',
-  'Californium',
-  'Einsteinium',
-  'Fermium',
-  'Mendelevium',
-  'Nobelium',
-  'Lawrencium',
-  'Rutherfordium',
-  'Dubnium',
-  'Seaborgium',
-  'Bohrium',
-  'Hassium',
-  'Meitnerium',
-  'Darmstadtium',
-  'Roentgenium',
-  'Copernicium',
-  'Nihonium',
-  'Flerovium',
-  'Moscovium',
-  'Livermorium',
-  'Tennessine',
-  'Oganesson',
-]
-
-const dataTestIds = {
-  toggleButton: 'toggle-button-id',
-  menu: 'menu-id',
-  item: index => `item-id-${index}`,
-  input: 'input-id',
-  combobox: 'combobox-id',
-}
-
-const DropdownCombobox = props => {
-  const {
-    isOpen,
-    getToggleButtonProps,
-    getLabelProps,
-    getMenuProps,
-    getInputProps,
-    getComboboxProps,
-    highlightedIndex,
-    getItemProps,
-  } = useCombobox({items, ...props})
-  return (
-    <div>
-      <label {...getLabelProps()}>Choose an element:</label>
-      <div data-testid={dataTestIds.combobox} {...getComboboxProps()}>
-        <input data-testid={dataTestIds.input} {...getInputProps()} />
-        <button
-          data-testid={dataTestIds.toggleButton}
-          {...getToggleButtonProps()}
-        >
-          Toggle
-        </button>
-      </div>
-      <ul data-testid={dataTestIds.menu} {...getMenuProps()}>
-        {isOpen &&
-          (props.items || items).map((item, index) => {
-            const stringItem =
-              item instanceof Object ? defaultProps.itemToString(item) : item
-            return (
-              <li
-                data-testid={dataTestIds.item(index)}
-                style={
-                  highlightedIndex === index ? {backgroundColor: 'blue'} : {}
-                }
-                key={`${stringItem}${index}`}
-                {...getItemProps({item, index})}
-              >
-                {stringItem}
-              </li>
-            )
-          })}
-      </ul>
-    </div>
-  )
-}
-
-const setup = props => render(<DropdownCombobox {...props} />)
-
-const defaultIds = getElementIds(useId)
-
-const setupHook = props => {
-  return renderHook(() => useCombobox({items, ...props}))
-}
-
-export {setupHook, dataTestIds, defaultIds, items, setup}
