+++ src/hooks/useSelect/__tests__/utils.test.js
@@ -1,7 +1,37 @@
-import {getItemIndexByCharacterKey} from '../utils'
+import {
+  getItemIndex,
+  getItemIndexByCharacterKey,
+  itemToString,
+} from '../../utils'
+import {getA11yStatusMessage} from '../utils'
 import reducer from '../reducer'
 
 describe('utils', () => {
+  describe('itemToString', () => {
+    test('returns empty string if item is falsy', () => {
+      const emptyString = itemToString(null)
+      expect(emptyString).toBe('')
+    })
+  })
+
+  describe('getItemIndex', () => {
+    test('returns -1 if no items', () => {
+      const index = getItemIndex(undefined, {}, [])
+      expect(index).toBe(-1)
+    })
+
+    test('returns index if passed', () => {
+      const index = getItemIndex(5, {}, [])
+      expect(index).toBe(5)
+    })
+
+    test('returns index of item', () => {
+      const item = {x: 2}
+      const index = getItemIndex(undefined, item, [{x: 1}, item, {x: 2}])
+      expect(index).toBe(1)
+    })
+  })
+
   describe('getItemIndexByCharacterKey', () => {
     const items = ['a', 'b', 'aba', 'aab', 'bab']
 
@@ -18,6 +48,18 @@ describe('utils', () => {
     test('checks from highlightedIndex position exclusively if there is only one key', () => {
       const index = getItemIndexByCharacterKey('a', 2, items, item => item)
       expect(index).toBe(3)
+    })
+  })
+
+  describe('getA11yStatusMessage', () => {
+    test('returns empty if no items', () => {
+      const message = getA11yStatusMessage({})
+      expect(message).toBe('')
+    })
+
+    test('returns empty if no message can be created', () => {
+      const message = getA11yStatusMessage({items: [], isOpen: false})
+      expect(message).toBe('')
     })
   })
