+++ src/index.js
@@ -1,3 +1,3 @@
 export {default} from './downshift'
 export {resetIdCounter} from './utils'
-export {useSelect, useCombobox} from './hooks'
+export {useSelect} from './hooks'
