+++ test/basic.test.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react'
-import Downshift, {StateChangeOptions} from '../'
+import Downshift, {StateChangeOptions, DownshiftInterface} from '../'
 
 type Item = string
+const TypedDownShift: DownshiftInterface<Item> = Downshift
 
 interface Props {}
 
@@ -26,7 +27,7 @@ export default class App extends React.Component<Props, State> {
     const items = this.state.items
 
     return (
-      <Downshift onChange={this.onChange}>
+      <TypedDownShift onChange={this.onChange}>
         {({
           getToggleButtonProps,
           getInputProps,
@@ -68,7 +69,7 @@ export default class App extends React.Component<Props, State> {
             ) : null}
           </div>
         )}
-      </Downshift>
+      </TypedDownShift>
     )
   }
 }
